pyproject.toml
[tool.poetry]
name = "shooting club"
version = "0.1.0"
description = ""
authors = ["wynardtage0 <fishface6666@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
aiogram = "^3.6.0"
SQLAlchemy = "^2.0.30"
python-dotenv = "^1.0.1"
redis = "^5.0.4"
pydantic = "^2.7.1"
asyncpg = "^0.29.0"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
